// SDKs/twilio.com/twilio_ip_messaging_v2/1.37.3/typescript/src/sdk/models/shared/ipmessagingv2serviceuser.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class IpMessagingV2ServiceUser extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=attributes" })
  attributes?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @SpeakeasyMetadata({ data: "json, name=identity" })
  identity?: string;

  @SpeakeasyMetadata({ data: "json, name=is_notifiable" })
  isNotifiable?: boolean;

  @SpeakeasyMetadata({ data: "json, name=is_online" })
  isOnline?: boolean;

  @SpeakeasyMetadata({ data: "json, name=joined_channels_count" })
  joinedChannelsCount?: number;

  @SpeakeasyMetadata({ data: "json, name=links" })
  links?: Map<string, any>;

  @SpeakeasyMetadata({ data: "json, name=role_sid" })
  roleSid?: string;

  @SpeakeasyMetadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}